refactor: 입력 뷰 인터페이스화 및 콘솔 입력 뷰 구현
